Update combat to change target when its dead
Sometimes the bot can get stuck on a dead target and get beaten by another mob that got pulled. Right now the plan switches to NO PLAN as the combat doesn't happen if the target is dead. 

This makes sure that if we are in combat and target is dead we clear target and change to the next target nearby. Especially useful if the bot pulled multiple mobs